1.9.1 Refactored components, added wizard component
