+++ src/frontend/src/BookingApp.sass
@@ -37,28 +37,20 @@
     justify-content: center
     font-size: calc(10px + 2vmin)
     color: black
-    
-  .Confirm
-    background-color: white
-    padding-right: 50px
-    padding-left: 50px
-    padding-top: 30px
-    padding-bottom: 20px
-    border: 1px solid #E0E0E0
-    opacity: .9
-    width: 80%
 
 
-  .AdditionalInfo
+  .text_box
     background-color: white
-    width: 80%
     margin-top: 10px
+    margin: 0 auto
     padding-right: 50px
     padding-left: 50px
     padding-top: 30px
     padding-bottom: 20px
     border: 1px solid #E0E0E0
     opacity: .9
+    min-width: 30vw
+    border-radius: 25px
